Add files via upload
Abstract:
These days, people's opinions and product reviews are quite important. These are disseminated via media outlets such as YouTube, personal blogs, and social media 
sites. Websites like Facebook and Twitter play an important role. Twitter is one such site, which combines microblogging with social networking. Tweets allow us to 
communicate our thoughts and opinions on Twitter. The major goal of our project is to capture live data and analyze it in real time. Sentiment analysis on live social media's mass data on targeted events, allowing the user to immediately see the impact of his statement on many industries such as the financial market, equities, and tech industry., so that the user can use this report to make last-minute changes to the event in order to change people's minds about it. This application is also utilized to get the most out of day stock trading.